+++ src/main/java/com/recipe/app/src/appVersion/infra/AppVersionCustomRepository.java
@@ -1,0 +1,8 @@
+package com.recipe.app.src.appVersion.infra;
+
+import com.recipe.app.src.appVersion.domain.AppVersion;
+
+public interface AppVersionCustomRepository {
+
+    AppVersion findRecentAppVersion();
+}
