+++ src/main/java/com/recipe/app/src/appVersion/infra/AppVersionRepository.java
@@ -1,0 +1,7 @@
+package com.recipe.app.src.appVersion.infra;
+
+import com.recipe.app.src.appVersion.domain.AppVersion;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AppVersionRepository extends JpaRepository<AppVersion, Long>, AppVersionCustomRepository {
+}
