+++ src/main/java/com/recipe/app/src/common/exception/BadWordException.java
@@ -1,0 +1,8 @@
+package com.recipe.app.src.common.exception;
+
+public class BadWordException extends RuntimeException {
+
+    public BadWordException(String word) {
+        super(String.format("금칙어 설정된 단어(%s)를 포함하고 있습니다.", word));
+    }
+}
