+++ src/main/java/com/recipe/app/src/dialog/application/DialogService.java
@@ -1,0 +1,23 @@
+package com.recipe.app.src.dialog.application;
+
+import com.recipe.app.src.dialog.application.dto.DialogResponse;
+import com.recipe.app.src.notice.exception.NotFoundNoticeException;
+import com.recipe.app.src.dialog.infra.DialogRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class DialogService {
+    private final DialogRepository dialogRepository;
+
+    public DialogService(DialogRepository dialogRepository) {
+        this.dialogRepository = dialogRepository;
+    }
+
+    @Transactional(readOnly = true)
+    public DialogResponse findDialog() {
+
+        return DialogResponse.from(dialogRepository.findFirstByActiveYn("Y")
+                .orElseThrow(NotFoundNoticeException::new));
+    }
+}
