--- src/main/java/com/recipe/app/src/etc/domain/AppVersion.java
@@ -1,26 +1,0 @@
-package com.recipe.app.src.etc.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "AppVersion")
-public class AppVersion {
-
-    @Id
-    @Column(name = "appVersionId", nullable = false, updatable = false)
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long appVersionId;
-
-    @Column(name = "version", nullable = false, length = 20)
-    private String version;
-}
