--- src/main/java/com/recipe/app/src/etc/infra/AppVersionCustomRepository.java
@@ -1,8 +1,0 @@
-package com.recipe.app.src.etc.infra;
-
-import com.recipe.app.src.etc.domain.AppVersion;
-
-public interface AppVersionCustomRepository {
-
-    AppVersion findRecentAppVersion();
-}
