+++ src/main/java/com/recipe/app/src/notice/domain/Notice.java
@@ -1,0 +1,33 @@
+package com.recipe.app.src.notice.domain;
+
+import com.recipe.app.src.common.entity.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+@Table(name = "Notice")
+public class Notice extends BaseEntity {
+
+    @Id
+    @Column(name = "noticeId", nullable = false, updatable = false)
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long noticeId;
+
+    @Column(name = "title", nullable = false, length = 128)
+    private String title;
+
+    @Column(name = "content", nullable = false, length = 2000)
+    private String content;
+
+    @Column(name = "activeYn", nullable = false, length = 1)
+    private String activeYn = "Y";
+}
